requirements.txt
biopython==1.78
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
dgl==0.9.1
et-xmlfile==1.1.0
future==0.18.3
idna @ file:///croot/idna_1666125576474/work
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.3.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
openpyxl==3.1.2
pandas==1.1.3
Pillow==9.5.0
psutil==5.9.5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2023.3
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn==0.24.1
scipy==1.5.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
threadpoolctl==3.1.0
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.1
typing_extensions==4.6.3
urllib3 @ file:///croot/urllib3_1673575502006/work
xlrd==1.2.0
